Extract and re-use calculateStatsPerFile
