+++ lib/eslint/eslint-helpers.js
@@ -664,6 +664,42 @@ function createIgnoreResult(filePath, baseDir, configStatus) {
 	};
 }
 
+/**
+ * It will calculate the error and warning count for collection of messages per file
+ * @param {LintMessage[]} messages Collection of messages
+ * @returns {Object} Contains the stats
+ * @private
+ */
+function calculateStatsPerFile(messages) {
+    const stat = {
+        errorCount: 0,
+        fatalErrorCount: 0,
+        warningCount: 0,
+        fixableErrorCount: 0,
+        fixableWarningCount: 0
+    };
+
+    for (let i = 0; i < messages.length; i++) {
+        const message = messages[i];
+
+        if (message.fatal || message.severity === 2) {
+            stat.errorCount++;
+            if (message.fatal) {
+                stat.fatalErrorCount++;
+            }
+            if (message.fix) {
+                stat.fixableErrorCount++;
+            }
+        } else {
+            stat.warningCount++;
+            if (message.fix) {
+                stat.fixableWarningCount++;
+            }
+        }
+    }
+    return stat;
+}
+
 //-----------------------------------------------------------------------------
 // Options-related Helpers
 //-----------------------------------------------------------------------------
@@ -988,8 +1024,9 @@ module.exports = {
 	isNonEmptyString,
 	isArrayOfNonEmptyString,
 
-	createIgnoreResult,
-	isErrorMessage,
+    createIgnoreResult,
+    isErrorMessage,
+    calculateStatsPerFile,
 
 	processOptions,
