+++ lib/eslint/eslint.js
@@ -31,8 +31,9 @@ const {
 	isNonEmptyString,
 	isArrayOfNonEmptyString,
 
-	createIgnoreResult,
-	isErrorMessage,
+    createIgnoreResult,
+    isErrorMessage,
+    calculateStatsPerFile,
 
 	processOptions,
 } = require("./eslint-helpers");
@@ -109,42 +110,6 @@ const removedFormatters = new Set([
 	"unix",
 	"visualstudio",
 ]);
-
-/**
- * It will calculate the error and warning count for collection of messages per file
- * @param {LintMessage[]} messages Collection of messages
- * @returns {Object} Contains the stats
- * @private
- */
-function calculateStatsPerFile(messages) {
-	const stat = {
-		errorCount: 0,
-		fatalErrorCount: 0,
-		warningCount: 0,
-		fixableErrorCount: 0,
-		fixableWarningCount: 0,
-	};
-
-	for (let i = 0; i < messages.length; i++) {
-		const message = messages[i];
-
-		if (message.fatal || message.severity === 2) {
-			stat.errorCount++;
-			if (message.fatal) {
-				stat.fatalErrorCount++;
-			}
-			if (message.fix) {
-				stat.fixableErrorCount++;
-			}
-		} else {
-			stat.warningCount++;
-			if (message.fix) {
-				stat.fixableWarningCount++;
-			}
-		}
-	}
-	return stat;
-}
 
 /**
  * Create rulesMeta object.
