+++ lib/services/suppressed-violations-manager.js
@@ -6,6 +6,7 @@
 
 const fs = require("node:fs");
 const path = require("node:path");
+const { calculateStatsPerFile } = require("../eslint/eslint-helpers");
 
 //-----------------------------------------------------------------------------
 // Exports
@@ -266,24 +267,8 @@ class SuppressedViolationsManager {
                 })
         );
 
-        suppressedMessages.forEach(message => {
-            if (message.fatal || message.severity === 2) {
-                result.errorCount--;
-                if (message.fatal) {
-                    result.fatalErrorCount--;
-                }
-                if (message.fix) {
-                    result.fixableErrorCount--;
-                }
-            } else {
-                result.warningCount--;
-                if (message.fix) {
-                    result.fixableWarningCount--;
-                }
-            }
-        });
-
         result.messages = result.messages.filter(message => message.ruleId !== ruleId);
+        Object.assign(result, calculateStatsPerFile(result.messages));
     }
 }
